requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
movielens-private-api==0.0.1
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
scikit-surprise==1.1.3
scipy==1.13.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
